+++ src/background/state/application/apps_config.rs
@@ -1,20 +1,20 @@
 use windows::Win32::Foundation::HWND;
 
-use crate::{state::domain::AppConfig, windows_api::window::Window};
+use crate::{error::Result, state::domain::AppConfig, windows_api::window::Window};
 
 use super::FullState;
 
 impl FullState {
-    pub fn get_app_config_by_window(&self, hwnd: HWND) -> Option<&AppConfig> {
+    pub fn get_app_config_by_window(&self, hwnd: HWND) -> Result<Option<&AppConfig>> {
         let window = Window::from(hwnd);
 
-        let path = window.process().program_path().ok()?;
+        let path = window.process().program_path()?;
 
-        let exe = path.file_name()?.to_string_lossy();
+        let exe = path.file_name().ok_or("Invalid path")?.to_string_lossy();
         let path = path.to_string_lossy();
         let title = window.title();
         let class = window.class();
 
-        self.settings_by_app.search(&title, &class, &exe, &path)
+        Ok(self.settings_by_app.search(&title, &class, &exe, &path))
     }
 }